uc-crux-llvm: Factor out type translation into a module
