+++ uc-crux-llvm/src/UCCrux/LLVM/Overrides/Skip.hs
@@ -70,7 +70,8 @@ import           UCCrux.LLVM.Constraints (ConstrainedShape(ConstrainedShape))
 import           UCCrux.LLVM.Context.Module (ModuleContext, declTypes, moduleTypes)
 import           UCCrux.LLVM.Cursor (Selector(SelectReturn), Cursor(Here))
 import           UCCrux.LLVM.Errors.Panic (panic)
-import           UCCrux.LLVM.FullType.CrucibleType (FunctionTypes, toCrucibleType, lookupDeclTypes, ftRetType, isDebug)
+import           UCCrux.LLVM.FullType.CrucibleType (toCrucibleType)
+import           UCCrux.LLVM.FullType.Translation (FunctionTypes, lookupDeclTypes, ftRetType, isDebug)
 import           UCCrux.LLVM.Setup (SymValue(getSymValue), generate)
 import           UCCrux.LLVM.Setup.Monad (TypedSelector, runSetup, resultAssumptions, resultMem, ppSetupError, resultAnnotations)
 import qualified UCCrux.LLVM.Shape as Shape
